+++ policyengine_canada/variables/gov/provinces/ab/benefits/acfb/acfb_eligible_children.py
@@ -1,0 +1,10 @@
+from policyengine_canada.model_api import *
+
+
+class acfb_eligible_children(Variable):
+    value_type = int
+    entity = Household
+    label = "Alberta child and family benefit eligible children"
+    definition_period = YEAR
+
+    adds = ["acfb_eligible_child"]